feature: SECOND INCREMENT FINISHED
The user is now able to register, log in and view his profile and others.

The app is fully connected to the back-end Supabase database via Pinia.

Components created in the first increment have been modified and improved.

Not all the criteria for this increment have been met since the time limit is approaching.