feature: THIRD INCREMENT FINISHED
The user is now able to upload items to the data base so others can see and cooperate in finding them.

Users(logged in or not) can notify a finding of an item clicking on the notify button.

Created basic filters of items.

Not all the criteria for this increment have been met since the time limit is approaching.

In the next increment I will develop USs related to the fifth increment, leaving for the future USs related to the fourth one since the time limit is approaching.